+++ src/components/password/password-test.stories.tsx
@@ -205,6 +205,23 @@ export const NewValidation = () => {
         warning="Warning Message"
         value={state}
         onChange={setValue}
+        mb={2}
+      />
+      <Password
+        validationMessagePositionTop={false}
+        label="Password"
+        error="Error Message"
+        inputHint="Hint text (optional)."
+        value={state}
+        onChange={setValue}
+        mb={2}
+      />
+      <Password
+        validationMessagePositionTop={false}
+        label="Password"
+        warning="Warning Message"
+        value={state}
+        onChange={setValue}
       />
     </CarbonProvider>
   );