+++ src/components/textarea/textarea-test.stories.tsx
@@ -203,7 +203,18 @@ export const NewValidation = () => {
   return (
     <CarbonProvider validationRedesignOptIn>
       <Textarea label="Textarea" error="Error Message" mb={2} />
-      <Textarea label="Textarea" warning="Warning Message" />
+      <Textarea label="Textarea" warning="Warning Message" mb={2} />
+      <Textarea
+        validationMessagePositionTop={false}
+        label="Textarea"
+        error="Error Message"
+        mb={2}
+      />
+      <Textarea
+        validationMessagePositionTop={false}
+        label="Textarea"
+        warning="Warning Message"
+      />
     </CarbonProvider>
   );
 };